A new ingester command
This command allows a folder (and its sub-folders) on a local device to
be duplicated in Indigo.  The command is ...

    indigo ingest --group Administrators --user ross  --folder .

The local directory structure will be replicated in Indigo with all
files found (apart from those extensions found in the SKIP list) being
added as resources to the collection corresponding to the local
directory.

Write, Edit and Delete permissions are given to the group specified on
the command line, and the user provided is simply for recording activity